+++ crates/apollo-compiler/test_data/err/0008_operation_with_undefined_variables_in_inline_fragment.graphql
@@ -1,4 +1,4 @@
-query ExampleQuery($variable: Int) {
+query ExampleQuery {
   topProducts {
     name
   }
