+++ crates/apollo-compiler/test_data/err/0008_operation_with_undefined_variables_in_inline_fragment.txt
@@ -1,1 +1,1 @@
-[Error(UndefinedVariable { message: "Variable undefined", variable: "value" }), Warning(UnusedVariable { message: "unused variable", variable: "variable" })]
+[Error(UndefinedVariable { message: "Variable undefined", variable: "value" })]
