+++ crates/apollo-compiler/test_data/err/0010_operation_with_unused_variable.graphql
@@ -1,0 +1,6 @@
+query ExampleQuery($unusedVariable: Int) {
+  topProducts {
+    name
+  }
+  ... multipleSubscriptions
+}
