+++ crates/apollo-compiler/test_data/err/0010_operation_with_unused_variable.txt
@@ -1,0 +1,1 @@
+[Warning(UnusedVariable { message: "unused variable", variable: "unusedVariable" })]
