+++ crates/apollo-compiler/test_data/err/0011_syntactic_errors_from_parser.graphql
@@ -1,0 +1,6 @@
+query getName {}
+
+fragment fragmentOne {
+    name
+    birthday
+}
